+++ DieLemmaDressBlock.py
@@ -1,9 +1,9 @@
 # This is a pattern Jane made when we were working on The Chicken Dress for Die Lemma (August 2022)
 # I'm digitising it here based on measurements taken from the block
 
-from geometry.Vector import Vector
-from geometry.Shape import Shape
 from render import render
+from src.geometry.Shape import Shape
+from src.geometry.Vector import Vector
 
 inch = 25.4
