+++ TheClassicTailoredTrouserBlock.py
@@ -1,7 +1,8 @@
-from geometry.Vector import Vector, midpoint
-from geometry.Shape import Shape
 from BodyMeasurements import exampleBodyMeasurements, BodyMeasurements
-from geometry.Group import Group
+from src.geometry.Group import Group
+from src.geometry.Shape import Shape
+from src.geometry.Vector import midpoint, Vector
+
 
 def TheClassicTailoredTrouserBlock(body: BodyMeasurements = exampleBodyMeasurements, bottomWidth: float = 220):
     "Create a trouser block according to Winnifred Owen instructions"
