--- requirements.txt
@@ -1,2 +1,0 @@
-numpy==1.23.3
-drawsvg==1.8.3
