+++ src/geometry/Group.py
@@ -1,5 +1,5 @@
 import drawSvg as svg
-from geometry.isMovable import isMovable
+from src.geometry.isMovable import isMovable
 
 class Group:
     label = None
