+++ src/geometry/Line.py
@@ -1,4 +1,4 @@
-from geometry.Vector import Vector
+from src.geometry.Vector import Vector
 
 class StraightLine:
     def __init__(self, gradient, intercept):
