+++ src/geometry/Shape3d.py
@@ -1,6 +1,5 @@
-from geometry.Vector import Vector
-from geometry.vec3 import vec3
-from geometry.Shape import Shape
+from src.geometry.Vector import Vector
+from src.geometry.Shape import Shape
 
 class Shape3d:
