+++ src/geometry/vec3.py
